% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundancemodelsJags.R
\name{countmix}
\alias{countmix}
\title{Hierarchical Bayesian models'}
\usage{
countmix(count, mgmt, hab, source, species = c("YCHUB", "BSHINER"))
}
\arguments{
\item{count}{Data frame with count data including pond_name, year, site, YC_small, BS_small}

\item{mgmt}{Data frame with management data including pond_name, year, and stocking/removal variables}

\item{hab}{Data frame with habitat variables including pond_name, year, site, pH, wtemp, doxygen, wcond, veg, wdepth}

\item{species}{Character string: "YCHUB" or "BSHINER"}
}
\description{
Hierarchical Bayesian models'
}
