% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifiedrandom.R
\name{stratified}
\alias{stratified}
\title{Stratified Random Sampling R Function'}
\usage{
stratified(df, group, size, select = NULL, replace = FALSE, bothSets = FALSE)
}
\description{
Stratified Random Sampling R Function'
}
